requirements.txt
# Backend dependencies
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy
faiss-cpu
alembic==1.13.0
python-multipart==0.0.6


# Common dependencies
python-dotenv==1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
sentence-transformers

# Together AI dependencies
together>=0.2.5

# Frontend dependencies
# streamlit>=1.24.0
requests>=2.31.0
